backend/requirements.txt
fastapi
uvicorn
discord.py
apscheduler
sqlalchemy
pydantic
